remove braintrust scoring_fn
